move task; delete unused models
